load multiple workspace folders into the same project
This is a hack to allow cquery to work with a workspace that contains
multiple project in different workspace folders. It just aggregates
all the entries and include directories from all the workspace folders
into a single project and index cache.

I think the correct solution would be to have a different Project for
each workspace folder, but that would require changes in the code in
several locations.

This hack also doesn't handle the addition or removal of workspace
folders to the workspace (https://microsoft.github.io/language-server-protocol/specification#workspace_didChangeWorkspaceFolders)